// src/api/generated/Preferences.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Theme } from "./Theme";

export type Preferences = { theme: Theme, idle_lock_seconds: number, minimize_to_tray: boolean, launch_on_startup: boolean, notify_no_payment_id: boolean, notify_weak_password: boolean, };