// src/api/generated/NodeOptions.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type NodeOptions = { daemon_log_level: number, wallet_log_level: number, in_peers: number, out_peers: number, limit_rate_up: number, limit_rate_down: number, p2p_port: number, rpc_port: number, zmq_port: number, wallet_rpc_port: number, public_p2p: boolean, };